Revert "feat(core): move runtime-lint-utils to eslint plugin (#13222)"
This reverts commit ab6f62a9b5186ea96aa409a6dbe73586f550da71.
