+++ packages/cypress/plugins/cypress-preset.ts
@@ -8,8 +8,8 @@ import {
   TargetConfiguration,
   workspaceRoot,
 } from '@nrwl/devkit';
+import { mapProjectGraphFiles } from '@nrwl/workspace/src/utils/runtime-lint-utils';
 import { readProjectsConfigurationFromProjectGraph } from 'nx/src/project-graph/project-graph';
-import { mapProjectGraphFiles } from 'nx/src/utils/target-project-locator';
 import { dirname, extname, join, relative } from 'path';
 import { lstatSync } from 'fs';
 
@@ -98,8 +98,8 @@ export function getProjectConfigByPath(
     !graph.nodes[componentTestingProjectName]?.data
   ) {
     throw new Error(
-      stripIndents`Unable to find the project configuration that includes ${normalizedPathFromWorkspaceRoot}.
-      Found project name? ${componentTestingProjectName}.
+      stripIndents`Unable to find the project configuration that includes ${normalizedPathFromWorkspaceRoot}. 
+      Found project name? ${componentTestingProjectName}. 
       Graph has data? ${!!graph.nodes[componentTestingProjectName]?.data}`
     );
   }
