+++ packages/eslint-plugin-nx/.eslintrc.json
@@ -3,6 +3,7 @@
   "rules": {
     "no-restricted-imports": [
       "error",
+      "@nrwl/workspace",
       "@angular-devkit/core",
       "@angular-devkit/architect",
       "@angular-devkit/schematics"
