+++ packages/eslint-plugin-nx/src/rules/enforce-module-boundaries.spec.ts
@@ -1,15 +1,13 @@
 import type { FileData, ProjectGraph } from '@nrwl/devkit';
 import { DependencyType } from '@nrwl/devkit';
+import { mapProjectGraphFiles } from '@nrwl/workspace/src/utils/runtime-lint-utils';
 import * as parser from '@typescript-eslint/parser';
 import { TSESLint } from '@typescript-eslint/utils';
 import { vol } from 'memfs';
-import {
-  TargetProjectLocator,
-  mapProjectGraphFiles,
-} from 'nx/src/utils/target-project-locator';
+import { TargetProjectLocator } from 'nx/src/utils/target-project-locator';
 import enforceModuleBoundaries, {
   RULE_NAME as enforceModuleBoundariesRuleName,
-} from './enforce-module-boundaries';
+} from '../../src/rules/enforce-module-boundaries';
 
 jest.mock('fs', () => require('memfs').fs);
