+++ packages/react/src/utils/ast-utils.ts
@@ -1,5 +1,5 @@
 import * as ts from 'typescript';
-import { findNodes } from 'nx/src/utils/typescript';
+import { findNodes } from '@nrwl/workspace/src/utilities/typescript/find-nodes';
 import {
   ChangeType,
   logger,
