+++ packages/workspace/src/utils/ast-utils.ts
@@ -18,12 +18,17 @@ import {
   Tree,
 } from '@angular-devkit/schematics';
 import * as ts from 'typescript';
-import { parseJson, serializeJson, FileData } from '@nrwl/devkit';
+import {
+  parseJson,
+  ProjectConfiguration,
+  serializeJson,
+  FileData,
+} from '@nrwl/devkit';
 import { getWorkspacePath } from './cli-config-utils';
 import { extname, join, normalize, Path } from '@angular-devkit/core';
-import type { NxJsonConfiguration } from '@nrwl/devkit';
+import type { NxJsonConfiguration, ProjectsConfigurations } from '@nrwl/devkit';
 import { addInstallTask } from './rules/add-install-task';
-import { findNodes } from 'nx/src/utils/typescript';
+import { findNodes } from '../utilities/typescript/find-nodes';
 import { getSourceNodes } from '../utilities/typescript/get-source-nodes';
 
 function nodesByPosition(first: ts.Node, second: ts.Node): number {
@@ -66,6 +71,7 @@ export function sortObjectByKeys(obj: unknown) {
     }, {});
 }
 
+export { findNodes } from '../utilities/typescript/find-nodes';
 export { getSourceNodes } from '../utilities/typescript/get-source-nodes';
 
 export interface Change {
