+++ packages/workspace/src/utils/graph-utils.ts
