+++ packages/workspace/src/utils/rules/rename-package-imports.ts
@@ -7,8 +7,7 @@ import {
 } from '@angular-devkit/schematics';
 import { getWorkspace } from '../workspace';
 import { visitNotIgnoredFiles } from './visit-not-ignored-files';
-import { insert, ReplaceChange } from '../ast-utils';
-import { findNodes } from 'nx/src/utils/typescript';
+import { findNodes, insert, ReplaceChange } from '../ast-utils';
 import { normalize } from '@angular-devkit/core';
 
 export interface PackageNameMapping {
